ingestion_lambda_function/requirements.txt
#boto3
qdrant-client
google-genai
mangum
fastapi
langchain-text-splitters
python-multipart
uvicorn
